refactor(utils/pause/frames) - Finished putting things in their places
(#5616)
