+++ src/utils/pause/frames/getFrameUrl.js
@@ -1,0 +1,13 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at <http://mozilla.org/MPL/2.0/>. */
+
+// @flow
+
+import { get } from "lodash";
+
+import type { Frame } from "../../../types";
+
+export function getFrameUrl(frame: Frame) {
+  return get(frame, "source.url", "") || "";
+}
