+++ src/utils/pause/scopes/getScope.js
@@ -5,7 +5,7 @@
 // @flow
 import { getBindingVariables } from "./getVariables";
 import { getFramePopVariables, getThisVariable } from "./utils";
-import { simplifyDisplayName } from "../../frame";
+import { simplifyDisplayName } from "../../pause/frames";
 
 import type { Frame, Why, Scope, BindingContents } from "../../../types";
